refactor utils::iter
