+++ src/utils/iter/lazy_sort.rs
@@ -1,132 +1,4 @@
-use std::{cmp::Ordering, collections::BinaryHeap, fmt, iter::Peekable, ops::Deref};
-
-use dyn_clone::{clone_trait_object, DynClone};
-
-/// Wrapper around an iterator with additional abilities like cloning.
-pub struct LazyList<T>(pub(crate) Box<dyn CloneableIterator<Item = T>>);
-
-impl<T> Clone for LazyList<T> {
-    fn clone(&self) -> Self {
-        Self(self.0.clone())
-    }
-}
-
-impl<T> PartialEq for LazyList<T> {
-    fn eq(&self, _other: &Self) -> bool {
-        // TODO: define rules for partial equality
-        false
-    }
-}
-
-impl<T> Eq for LazyList<T> {}
-
-impl<T> PartialOrd for LazyList<T> {
-    fn partial_cmp(&self, _other: &Self) -> Option<Ordering> {
-        // TODO: define rules for partial cmp
-        None
-    }
-}
-
-impl<T> Deref for LazyList<T> {
-    type Target = Box<dyn CloneableIterator<Item = T>>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl<T> fmt::Debug for LazyList<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let (lower_bound, upper_bound) = self.size_hint();
-        let size: Box<dyn fmt::Debug> = if upper_bound == Some(lower_bound) {
-            Box::new(lower_bound)
-        } else if let Some(upper) = upper_bound {
-            Box::new(lower_bound..=upper)
-        } else {
-            Box::new(lower_bound..)
-        };
-
-        let name = format!("LazyList<{}>", std::any::type_name::<T>());
-        f.debug_struct(&name).field("size", &size).finish()
-    }
-}
-
-impl<T> Iterator for LazyList<T> {
-    type Item = T;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.0.next()
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.0.size_hint()
-    }
-}
-
-impl<T> LazyList<T> {
-    pub(crate) fn extend<I>(&mut self, iter: I)
-    where
-        T: 'static,
-        I: IntoIterator<Item = T>,
-        I::IntoIter: Clone + 'static,
-    {
-        let mut content: Box<dyn CloneableIterator<Item = T>> = Box::new(std::iter::empty());
-        std::mem::swap(&mut self.0, &mut content);
-        self.0 = Box::new(content.chain(iter));
-    }
-}
-
-/// Clone-able [Iterator] which can be used in dyn context.
-pub trait CloneableIterator: Iterator + DynClone {}
-
-impl<I: Iterator + DynClone> CloneableIterator for I {}
-
-clone_trait_object!(<T> CloneableIterator<Item = T>);
-
-pub const fn append_with_last<I, F>(iter: I, f: F) -> AppendWithLast<I, F>
-where
-    I: Iterator,
-    I::Item: Clone,
-    F: FnMut(I::Item) -> Option<I::Item>,
-{
-    AppendWithLast {
-        iter,
-        last_item: None,
-        f,
-    }
-}
-
-pub struct AppendWithLast<I, F>
-where
-    I: Iterator,
-{
-    iter: I,
-    last_item: Option<I::Item>,
-    f: F,
-}
-
-impl<I, F> Iterator for AppendWithLast<I, F>
-where
-    I: Iterator,
-    I::Item: Clone,
-    F: FnMut(I::Item) -> Option<I::Item>,
-{
-    type Item = I::Item;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if let Some(item) = self.iter.next() {
-            self.last_item = Some(item.clone());
-            Some(item)
-        } else {
-            self.last_item.take().and_then(|last| (self.f)(last))
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let (lo, hi) = self.iter.size_hint();
-        (lo, hi.map(|hi| hi + 1))
-    }
-}
+use std::{cmp::Ordering, collections::BinaryHeap, iter::Peekable};
 
 /// Produce sorted iterator by zipping the iterator of pairs `(a, b)`.
 ///
@@ -166,7 +38,7 @@ type IsFirstFn<T> = Box<dyn Fn(&T, &T) -> bool>;
 
 struct OrdFromKeyWrapper<T> {
     item: T,
-    less_fn: IsFirstFn<T>
+    less_fn: IsFirstFn<T>,
 }
 
 impl<T> PartialEq for OrdFromKeyWrapper<T> {
@@ -253,27 +125,6 @@ where
         let (lo, hi) = self.iter.size_hint();
         (lo * 2, hi.map(|hi| hi * 2))
     }
-}
-
-pub fn partition<I, T, F, TakeF>(
-    iter: I,
-    predicate: F,
-    take_only: Option<TakeF>,
-) -> (
-    impl CloneableIterator<Item = T>,
-    impl CloneableIterator<Item = T>,
-)
-where
-    I: Iterator<Item = T> + Clone,
-    F: Fn(&T) -> bool + Clone + 'static,
-    TakeF: Fn(&T) -> bool + Clone + 'static,
-{
-    let take = move |x: &T| take_only.as_ref().map_or(true, |take_only| take_only(x));
-    let filter = move |x: &T| predicate(x);
-
-    let left = iter.clone().take_while(take.clone()).filter(filter.clone());
-    let right = iter.take_while(take).filter(move |x| !filter(x));
-    (left, right)
 }
 
 #[cfg(test)]
